Murano 2015.1.0rc1

+++ murano/db/catalog/api.py
@@ -14,6 +14,7 @@
 
 from oslo.config import cfg
 from oslo.db.sqlalchemy import utils
+import sqlalchemy as sa
 from sqlalchemy import or_
 from sqlalchemy.orm import attributes
 # TODO(ruhe) use exception declared in openstack/common/db
@@ -21,7 +22,7 @@ from webob import exc
 
 from murano.db import models
 from murano.db import session as db_session
-from murano.openstack.common.gettextutils import _
+from murano.common.i18n import _, _LW
 from murano.openstack.common import log as logging
 
 CONF = cfg.CONF
@@ -54,26 +55,24 @@ def _package_get(package_id_or_name, session):
         msg = _("Package id or name '{0}' not found").\
             format(package_id_or_name)
         LOG.error(msg)
-        raise exc.HTTPNotFound(msg)
+        raise exc.HTTPNotFound(explanation=msg)
 
     return package
 
 
 def _authorize_package(package, context, allow_public=False):
-    if context.is_admin:
-        return
 
     if package.owner_id != context.tenant:
         if not allow_public:
             msg = _("Package '{0}' is not owned by "
                     "tenant '{1}'").format(package.id, context.tenant)
             LOG.error(msg)
-            raise exc.HTTPForbidden(msg)
+            raise exc.HTTPForbidden(explanation=msg)
         if not package.is_public:
             msg = _("Package '{0}' is not public and not owned by "
                     "tenant '{1}' ").format(package.id, context.tenant)
             LOG.error(msg)
-            raise exc.HTTPForbidden(msg)
+            raise exc.HTTPForbidden(explanation=msg)
 
 
 def package_get(package_id_or_name, context):
@@ -83,7 +82,8 @@ def package_get(package_id_or_name, context):
     """
     session = db_session.get_session()
     package = _package_get(package_id_or_name, session)
-    _authorize_package(package, context, allow_public=True)
+    if not context.is_admin:
+        _authorize_package(package, context, allow_public=True)
     return package
 
 
@@ -104,7 +104,7 @@ def _get_categories(category_names, session=None):
             msg = _("Category '{name}' doesn't exist").format(name=ctg_name)
             LOG.error(msg)
             # it's not allowed to specify non-existent categories
-            raise exc.HTTPBadRequest(msg)
+            raise exc.HTTPBadRequest(explanation=msg)
 
         categories.append(ctg_obj)
     return categories
@@ -174,8 +174,8 @@ def _do_add(package, change):
         try:
             getattr(package, path).append(item)
         except AssertionError:
-            msg = _('One of the specified {0} is already '
-                    'associated with a package. Doing nothing.')
+            msg = _LW('One of the specified {0} is already '
+                      'associated with a package. Doing nothing.')
             LOG.warning(msg.format(path))
     return package
 
@@ -196,10 +196,31 @@ def _do_remove(package, change):
             msg = _("Value '{0}' of property '{1}' "
                     "does not exist.").format(value, path)
             LOG.error(msg)
-            raise exc.HTTPNotFound(msg)
+            raise exc.HTTPNotFound(explanation=msg)
         item_to_remove = find(current_values, lambda i: i.name == value)
         current_values.remove(item_to_remove)
     return package
+
+
+def _get_packages_for_category(session, category_id):
+    """Return detailed list of packages, belonging to the provided category
+    :param session:
+    :param category_id:
+    :return: list of dictionaries, containing id, name and package frn
+    """
+    pkg = models.Package
+    packages = (session.query(pkg.id, pkg.name, pkg.fully_qualified_name)
+                .filter(pkg.categories
+                        .any(models.Category.id == category_id))
+                .all())
+
+    result_packages = []
+    for package in packages:
+        id, name, fqn = package
+        result_packages.append({'id': id,
+                                'name': name,
+                                'fully_qualified_name': fqn})
+    return result_packages
 
 
 def package_update(pkg_id_or_name, changes, context):
@@ -214,10 +235,22 @@ def package_update(pkg_id_or_name, changes, context):
     session = db_session.get_session()
     with session.begin():
         pkg = _package_get(pkg_id_or_name, session)
-        _authorize_package(pkg, context)
+        was_private = not pkg.is_public
+        if not context.is_admin:
+            _authorize_package(pkg, context)
 
         for change in changes:
             pkg = operation_methods[change['op']](pkg, change)
+        became_public = pkg.is_public
+        class_names = [clazz.name for clazz in pkg.class_definitions]
+        if was_private and became_public:
+            with db_session.get_lock("public_packages", session):
+                _check_for_existing_classes(session, class_names, None,
+                                            check_public=True,
+                                            ignore_package_with_id=pkg.id)
+                _check_for_public_packages_with_fqn(session,
+                                                    pkg.fully_qualified_name,
+                                                    pkg.id)
         session.add(pkg)
     return pkg
 
@@ -249,7 +282,7 @@ def package_search(filters, context, limit=None):
 
     if context.is_admin:
         if not include_disabled:
-            #NOTE(efedorova): is needed for SA 0.7.9, but could be done
+            # NOTE(efedorova): is needed for SA 0.7.9, but could be done
             # simpler in SA 0.8. See http://goo.gl/9stlKu for a details
             query = session.query(pkg).filter(pkg.__table__.c.enabled)
         else:
@@ -308,7 +341,8 @@ def package_search(filters, context, limit=None):
                         condition = getattr(pkg, attr).any(
                             getattr(models, fk_fields[attr]).name.like(_word))
                         conditions.append(condition)
-                    else:
+                    elif isinstance(getattr(pkg, attr)
+                                    .property.columns[0].type, sa.String):
                         conditions.append(getattr(pkg, attr).like(_word))
         query = query.filter(or_(*conditions))
 
@@ -336,16 +370,37 @@ def package_upload(values, tenant_id):
     composite_attr_to_func = {'categories': _get_categories,
                               'tags': _get_tags,
                               'class_definitions': _get_class_definitions}
-    with session.begin():
+    is_public = values.get('is_public', False)
+
+    if is_public:
+        public_lock = db_session.get_lock("public_packages", session)
+    else:
+        public_lock = None
+    tenant_lock = db_session.get_lock("classes_of_" + tenant_id, session)
+    try:
+        _check_for_existing_classes(session, values.get('class_definitions'),
+                                    tenant_id, check_public=is_public)
+        if is_public:
+            _check_for_public_packages_with_fqn(
+                session,
+                values.get('fully_qualified_name'))
         for attr, func in composite_attr_to_func.iteritems():
             if values.get(attr):
                 result = func(values[attr], session)
                 setattr(package, attr, result)
                 del values[attr]
-
         package.update(values)
         package.owner_id = tenant_id
         package.save(session)
+        tenant_lock.commit()
+        if public_lock is not None:
+            public_lock.commit()
+    except Exception:
+        tenant_lock.rollback()
+        if public_lock is not None:
+            public_lock.rollback()
+        raise
+
     return package
 
 
@@ -355,8 +410,29 @@ def package_delete(package_id_or_name, context):
 
     with session.begin():
         package = _package_get(package_id_or_name, session)
-        _authorize_package(package, context)
+        if not context.is_admin and package.owner_id != context.tenant:
+            raise exc.HTTPForbidden(
+                explanation='Package is not owned by the'
+                            ' tenant "{0}"'.format(context.tenant))
         session.delete(package)
+
+
+def category_get(category_id, session=None, packages=False):
+    """Return category details
+       :param category_id: ID of a category, string
+       :returns: detailed information about category, dict
+    """
+    if not session:
+        session = db_session.get_session()
+
+    category = session.query(models.Category).get(category_id)
+    if not category:
+        msg = _("Category id or '{0}' not found").format(category_id)
+        LOG.error(msg)
+        raise exc.HTTPNotFound(msg)
+    if packages:
+        category.packages = _get_packages_for_category(session, category_id)
+    return category
 
 
 def categories_list():
@@ -379,6 +455,68 @@ def category_add(category_name):
 
     with session.begin():
         category.update({'name': category_name})
+        # NOTE(kzaitsev) update package_count, so we can safely access from
+        # outside the session
+        category.package_count = 0
         category.save(session)
 
     return category
+
+
+def category_delete(category_id):
+    """Delete a category by ID."""
+    session = db_session.get_session()
+
+    with session.begin():
+        category = session.query(models.Category).get(category_id)
+        if not category:
+            msg = _("Category id '{0}' not found").format(category_id)
+            LOG.error(msg)
+            raise exc.HTTPNotFound(msg)
+        session.delete(category)
+
+
+def _check_for_existing_classes(session, class_names, tenant_id,
+                                check_public=False,
+                                ignore_package_with_id=None):
+    if not class_names:
+        return
+    q = session.query(models.Class.name).filter(
+        models.Class.name.in_(class_names))
+    private_filter = None
+    public_filter = None
+    predicate = None
+    if tenant_id is not None:
+        private_filter = models.Class.package.has(
+            models.Package.owner_id == tenant_id)
+    if check_public:
+        public_filter = models.Class.package.has(
+            models.Package.is_public)
+    if private_filter is not None and public_filter is not None:
+        predicate = sa.or_(private_filter, public_filter)
+    elif private_filter is not None:
+        predicate = private_filter
+    elif public_filter is not None:
+        predicate = public_filter
+    if predicate is not None:
+        q = q.filter(predicate)
+    if ignore_package_with_id is not None:
+        q = q.filter(models.Class.package_id != ignore_package_with_id)
+    if q.first() is not None:
+        msg = _('Class with the same full name is already '
+                'registered in the visibility scope')
+        LOG.error(msg)
+        raise exc.HTTPConflict(msg)
+
+
+def _check_for_public_packages_with_fqn(session, fqn,
+                                        ignore_package_with_id=None):
+    q = session.query(models.Package.id).\
+        filter(models.Package.is_public).\
+        filter(models.Package.fully_qualified_name == fqn)
+    if ignore_package_with_id is not None:
+        q = q.filter(models.Package.id != ignore_package_with_id)
+    if q.first() is not None:
+        msg = _('Package with the same Name is already made public')
+        LOG.error(msg)
+        raise exc.HTTPConflict(msg)
